+++ lib/main.dart
@@ -3,9 +3,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:lojavirtual/firebase_options.dart';
-import 'package:lojavirtual/models/user_manager.dart';
-import 'package:lojavirtual/screen/screen.dart';
-import 'package:provider/provider.dart';
+import 'package:lojavirtual/src/screen/login_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -21,23 +19,20 @@ void main() async {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider(
-      create: (_) => UserManager(),
-      child: MaterialApp(
-        title: 'Loja Nicolau',
-        debugShowCheckedModeBanner: false,
-        theme: ThemeData(
-          colorScheme: ColorScheme.fromSwatch(
-            primarySwatch: Colors.indigo,
-          ).copyWith(secondary: const Color(0xff7d7543d8)),
-          appBarTheme: const AppBarTheme(
-            elevation: 0,
-          ),
-          visualDensity: VisualDensity.adaptivePlatformDensity,
-          scaffoldBackgroundColor: const Color(0xFF7D79D0),
+    return MaterialApp(
+      title: 'Loja Nicolau',
+      debugShowCheckedModeBanner: false,
+      theme: ThemeData(
+        colorScheme: ColorScheme.fromSwatch(
+          primarySwatch: Colors.indigo,
+        ).copyWith(secondary: const Color(0xff7d7543d8)),
+        appBarTheme: const AppBarTheme(
+          elevation: 0,
         ),
-        home: Screen(),
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+        scaffoldBackgroundColor: const Color(0xFF7D79D0),
       ),
+      home: LoginScreen(),
     );
   }
 }
