--- lib/models/user.dart
@@ -1,9 +1,0 @@
-class Userr {
-  Userr({
-    required this.email,
-    required this.password,
-  });
-
-  final String email;
-  final String password;
-}
