+++ lib/src/auth/components/custom_text_form_field.dart
@@ -9,7 +9,7 @@ class CustomFormTextField extends StatefulWidget {
     required this.label,
     required this.keyType,
     required this.controller,
-    this.enabled,
+    
     this.isDense,
     this.isSecret = false,
     required this.validator,
@@ -22,7 +22,7 @@ class CustomFormTextField extends StatefulWidget {
   final TextInputType keyType;
   final bool? isDense;
   final TextEditingController? controller;
-  final bool? enabled;
+
 
   @override
   State<CustomFormTextField> createState() => _CustomFormTextFieldState();
@@ -42,7 +42,6 @@ class _CustomFormTextFieldState extends State<CustomFormTextField> {
     return Padding(
       padding: const EdgeInsets.only(bottom: 15),
       child: TextFormField(
-        enabled: widget.enabled,
         controller: widget.controller,
         obscureText: isObscure,
         validator: widget.validator,
