+++ lib/src/auth/components/custom_text_logo.dart
